requirements.txt
streamlit
openai
langchain
python-dotenv
"unstructured[pdf, pptx]"